refactor: modularize and relocate connections handler to features directory
- Split and moved `load_admin_connections_handler` logic into `features/connections/whitelist.ts`.
- Streamlined event handling by simplifying `handleConnections` structure.
- Updated references in `events/index.ts` to reflect new file location.
