+++ behaviour_pack/scripts-dev/events/index.ts
@@ -1,6 +1,6 @@
 import load_block_event_handler from "../features/interactions/block-interact";
 import load_chat_handler from "../features/chat-decoration/chat";
-import load_connections_handler from "./connections";
+import load_connections_handler from "../features/connections/connections";
 import load_entity_event_handler from "../features/interactions/entity-die";
 import load_script_event_handler from "../features/interactions/script_events";
 import load_eliana_handler from "./eliana_trade";
